requirements.txt
face_recognition==1.3.0
face_recognition.egg==info
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==11.0.0
streamlit==1.39.0
streamlit.egg==info
